BUG#: 3178 TITLE: Fixed warning on Windows.
DESCRIPTION: Another 64-bit Windows issue.  Just added a static_cast.

+++ src/Pegasus/Common/tests/Tracer/Tracer.cpp
@@ -72,7 +72,7 @@ Uint32 compare(const char* fileName, const char* compareStr)
     Uint32 count=0;
     Uint32 retCode=0;
     fstream file;
-    Uint32 size=strlen(compareStr);
+    Uint32 size= static_cast<Uint32>(strlen(compareStr));
     AutoArrayPtr<char> readStr(new char[size+EOF_CHAR+1]);
 
     file.open(fileName,fstream::in);
